Redefine Transactable in terms of interface{} insetad of abci.Application
In go, if `A` implements `SomeInterface`, and `B` has an anonymous
member `*A`, `B` does _not_ implement `SomeInterface`. What this means
for us is that if apps are defined in terms of meta.App, the actual
application type no longer implements abci.Application. This then
causes interface conversion panics at runtime downstream.