Rework metastate into struct
We'd previously been treating metastate as a noms value and handling
all marshalling and unmarshalling ourselves. That didn't work out
for us, unfortunately, beacuse it led to a bunch of panics which
more or less made no sense.

By changing it into a struct and using the built-in marshalling,
we've eliminated the panics.

We still get a bunch of errors, but at least these make sense.