Add a State method to initialize the state
The chaosnode backing state keeps throwing nil pointer exceptions
on creation. I suspect that the fundamental reason this happens is
that the state contains an nt.Map object, which never actually
gets initialized with NewMap. There isn't a good place to do that
within the chaosnode NewApp function, which means that the best
solution is to add an interface method to State, and have it called
within meta.NewApp.

We'll see if I was right.