Move history inspection functions into metanode
The proximate cause of this move is that changing the semantics of
application height (creating noms blocks only when transactions exist,
instead of creating one per tendermint block) broke the chaos chain
badly enough that it made sense to refactor to fix things closer to
where they'd broken.

The ultimate cause of this move is that the ndau chain will eventually
also want to support history inspection, and it would be nonsense to
re-implement things from scratch there when we have a perfectly good
metanode sitting around from which both projects derive.

This appears to work correctly per the history tests in the testapp,
so it's time to try it on a real chain against tendermint.