Reduce the chaincode identifiers for round stats
These days, there is much less reason to use numbers other than a simple
sequence beginning at 0, because structs are no longer flattened within
the chaincode interpreter. When possible, we retain historical numbers
though, as practice: we never want to change something for the sake of it
if that change might not be backward-compatible.

In this case we have no choice: for now, chaincode identifiers must be
in the range 0-127. We therefore reduce these chain struct IDs to fit
into the legal range.

This enables us to write a chaincode script using these numbers in
https://github.com/oneiro-ndev/chaincode_scripts/pull/21 .